requirements.txt
ruff==0.11.13
pre_commit==4.2.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
python-dotenv==1.1.0